pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Make diff between two files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.1",
    "pytest-cov",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
